LayoutTests/fast/dom/microtask-inorder-expected.txt

CONSOLE MESSAGE: Finished queueing tests.
CONSOLE MESSAGE: MicroTask #2 has run.
CONSOLE MESSAGE: MicroTask #1 has run.
CONSOLE MESSAGE: MicroTask #0 has run.


--------
Frame: 'frame1'
--------


--------
Frame: 'frame2'
--------
